[flutter_tools] switch FakeCache to cache.test and NoopUsage to TestUsage (#76802)
