+++ packages/flutter_tools/test/commands.shard/hermetic/build_windows_test.dart
@@ -18,6 +18,7 @@ import 'package:process/process.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/fakes.dart';
 import '../../src/testbed.dart';
 
 const String flutterRoot = r'C:\flutter';
