+++ packages/flutter_tools/test/commands.shard/hermetic/doctor_test.dart
@@ -34,6 +34,7 @@ import 'package:fake_async/fake_async.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/fakes.dart';
 import '../../src/testbed.dart';
 
 final Generator _kNoColorOutputPlatform = () => FakePlatform(
