+++ packages/flutter_tools/test/commands.shard/hermetic/precache_test.dart
@@ -13,7 +13,7 @@ import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 void main() {
   MockCache cache;
