+++ packages/flutter_tools/test/commands.shard/permeable/devices_test.dart
@@ -20,6 +20,7 @@ import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
+import '../../src/fakes.dart';
 import '../../src/testbed.dart';
 
 void main() {
