+++ packages/flutter_tools/test/general.shard/emulator_test.dart
@@ -16,6 +16,7 @@ import 'package:mockito/mockito.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
+import '../src/fakes.dart';
 import '../src/mocks.dart';
 import '../src/testbed.dart';
