+++ packages/flutter_tools/test/general.shard/fuchsia/fuchsia_workflow_test.dart
@@ -11,7 +11,7 @@ import 'package:flutter_tools/src/fuchsia/fuchsia_sdk.dart';
 import 'package:flutter_tools/src/fuchsia/fuchsia_workflow.dart';
 
 import '../../src/common.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 void main() {
   final FileSystem fileSystem = MemoryFileSystem.test();
