+++ packages/flutter_tools/test/general.shard/linux/linux_device_test.dart
@@ -18,7 +18,7 @@ import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 final FakePlatform linux = FakePlatform(
   operatingSystem: 'linux',
