+++ packages/flutter_tools/test/general.shard/linux/linux_workflow_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter_tools/src/features.dart';
 import 'package:flutter_tools/src/linux/linux_workflow.dart';
 
 import '../../src/common.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 void main() {
   final Platform linux = FakePlatform(
