+++ packages/flutter_tools/test/general.shard/web/devices_test.dart
@@ -15,7 +15,7 @@ import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 void main() {
   testWithoutContext('No web devices listed if feature is disabled', () async {
