+++ packages/flutter_tools/test/general.shard/windows/windows_device_test.dart
@@ -18,7 +18,7 @@ import 'package:test/fake.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 void main() {
   testWithoutContext('WindowsDevice defaults', () async {
