+++ packages/flutter_tools/test/general.shard/windows/windows_workflow_test.dart
@@ -8,7 +8,7 @@ import 'package:flutter_tools/src/base/platform.dart';
 import 'package:flutter_tools/src/windows/windows_workflow.dart';
 
 import '../../src/common.dart';
-import '../../src/testbed.dart';
+import '../../src/fakes.dart';
 
 void main() {
   final FakePlatform windows = FakePlatform(operatingSystem: 'windows');
