+++ packages/flutter_tools/test/src/testbed.dart
@@ -19,7 +19,6 @@ import 'package:flutter_tools/src/base/terminal.dart';
 import 'package:flutter_tools/src/cache.dart';
 import 'package:flutter_tools/src/context_runner.dart';
 import 'package:flutter_tools/src/dart/pub.dart';
-import 'package:flutter_tools/src/features.dart';
 import 'package:flutter_tools/src/reporting/reporting.dart';
 import 'package:flutter_tools/src/version.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
@@ -29,6 +28,7 @@ import 'package:process/process.dart';
 import 'common.dart' as tester;
 import 'context.dart';
 import 'fake_http_client.dart';
+import 'fakes.dart';
 import 'throwing_pub.dart';
 
 export 'package:flutter_tools/src/base/context.dart' show Generator;
@@ -45,7 +45,7 @@ final Map<Type, Generator> _testbedDefaults = <Type, Generator>{
   ), // Allows reading logs and prevents stdout.
   OperatingSystemUtils: () => FakeOperatingSystemUtils(),
   OutputPreferences: () => OutputPreferences.test(), // configures BufferLogger to avoid color codes.
-  Usage: () => NoOpUsage(), // prevent addition of analytics from burdening test mocks
+  Usage: () => TestUsage(), // prevent addition of analytics from burdening test mocks
   FlutterVersion: () => FakeFlutterVersion(), // prevent requirement to mock git for test runner.
   Signals: () => FakeSignals(),  // prevent registering actual signal handlers.
   Pub: () => ThrowingPub(), // prevent accidental invocations of pub.
@@ -156,307 +156,4 @@ class Testbed {
       });
     }, createHttpClient: (SecurityContext c) => FakeHttpClient.any());
   }
-}
-
-/// A no-op implementation of [Usage] for testing.
-class NoOpUsage implements Usage {
-  @override
-  bool enabled = false;
-
-  @override
-  bool suppressAnalytics = true;
-
-  @override
-  String get clientId => 'test';
-
-  @override
-  Future<void> ensureAnalyticsSent() {
-    return null;
-  }
-
-  @override
-  Stream<Map<String, Object>> get onSend => const Stream<Map<String, Object>>.empty();
-
-  @override
-  void printWelcome() {}
-
-  @override
-  void sendCommand(String command, {Map<String, String> parameters}) {}
-
-  @override
-  void sendEvent(String category, String parameter, {
-    String label,
-    int value,
-    Map<String, String> parameters,
-  }) {}
-
-  @override
-  void sendException(dynamic exception) {}
-
-  @override
-  void sendTiming(String category, String variableName, Duration duration, { String label }) {}
-}
-
-class FakeFlutterVersion implements FlutterVersion {
-  @override
-  void fetchTagsAndUpdate() {  }
-
-  @override
-  String get channel => 'master';
-
-  @override
-  Future<void> checkFlutterVersionFreshness() async { }
-
-  @override
-  bool checkRevisionAncestry({String tentativeDescendantRevision, String tentativeAncestorRevision}) {
-    throw UnimplementedError();
-  }
-
-  @override
-  String get dartSdkVersion => '12';
-
-  @override
-  String get engineRevision => '42.2';
-
-  @override
-  String get engineRevisionShort => '42';
-
-  @override
-  Future<void> ensureVersionFile() async { }
-
-  @override
-  String get frameworkAge => null;
-
-  @override
-  String get frameworkCommitDate => null;
-
-  @override
-  String get frameworkDate => null;
-
-  @override
-  String get frameworkRevision => null;
-
-  @override
-  String get frameworkRevisionShort => null;
-
-  @override
-  String get frameworkVersion => null;
-
-  @override
-  GitTagVersion get gitTagVersion => null;
-
-  @override
-  String getBranchName({bool redactUnknownBranches = false}) {
-    return 'master';
-  }
-
-  @override
-  String getVersionString({bool redactUnknownBranches = false}) {
-    return 'v0.0.0';
-  }
-
-  @override
-  String get repositoryUrl => null;
-
-  @override
-  Map<String, Object> toJson() {
-    return null;
-  }
-}
-
-// A test implementation of [FeatureFlags] that allows enabling without reading
-// config. If not otherwise specified, all values default to false.
-class TestFeatureFlags implements FeatureFlags {
-  TestFeatureFlags({
-    this.isLinuxEnabled = false,
-    this.isMacOSEnabled = false,
-    this.isWebEnabled = false,
-    this.isWindowsEnabled = false,
-    this.isSingleWidgetReloadEnabled = false,
-    this.isAndroidEnabled = true,
-    this.isIOSEnabled = true,
-    this.isFuchsiaEnabled = false,
-    this.isExperimentalInvalidationStrategyEnabled = false,
-  });
-
-  @override
-  final bool isLinuxEnabled;
-
-  @override
-  final bool isMacOSEnabled;
-
-  @override
-  final bool isWebEnabled;
-
-  @override
-  final bool isWindowsEnabled;
-
-  @override
-  final bool isSingleWidgetReloadEnabled;
-
-  @override
-  final bool isAndroidEnabled;
-
-  @override
-  final bool isIOSEnabled;
-
-  @override
-  final bool isFuchsiaEnabled;
-
-  @override
-  final bool isExperimentalInvalidationStrategyEnabled;
-
-  @override
-  bool isEnabled(Feature feature) {
-    switch (feature) {
-      case flutterWebFeature:
-        return isWebEnabled;
-      case flutterLinuxDesktopFeature:
-        return isLinuxEnabled;
-      case flutterMacOSDesktopFeature:
-        return isMacOSEnabled;
-      case flutterWindowsDesktopFeature:
-        return isWindowsEnabled;
-      case singleWidgetReload:
-        return isSingleWidgetReloadEnabled;
-      case flutterAndroidFeature:
-        return isAndroidEnabled;
-      case flutterIOSFeature:
-        return isIOSEnabled;
-      case flutterFuchsiaFeature:
-        return isFuchsiaEnabled;
-      case experimentalInvalidationStrategy:
-        return isExperimentalInvalidationStrategyEnabled;
-    }
-    return false;
-  }
-}
-
-class FakeStatusLogger extends DelegatingLogger {
-  FakeStatusLogger(Logger delegate) : super(delegate);
-
-  Status status;
-
-  @override
-  Status startProgress(String message, {Duration timeout, String progressId, bool multilineOutput = false, int progressIndicatorPadding = kDefaultStatusPadding}) {
-    return status;
-  }
-}
-
-/// An implementation of the Cache which does not download or require locking.
-class FakeCache implements Cache {
-  @override
-  bool includeAllPlatforms;
-
-  @override
-  Set<String> platformOverrideArtifacts;
-
-  @override
-  bool useUnsignedMacBinaries;
-
-  @override
-  Future<bool> areRemoteArtifactsAvailable({String engineVersion, bool includeAllPlatforms = true}) async {
-    return true;
-  }
-
-  @override
-  String get dartSdkVersion => null;
-
-  @override
-  String get storageBaseUrl => null;
-
-  @override
-  MapEntry<String, String> get dyLdLibEntry => const MapEntry<String, String>('DYLD_LIBRARY_PATH', '');
-
-  @override
-  String get engineRevision => null;
-
-  @override
-  Directory getArtifactDirectory(String name) {
-    return globals.fs.currentDirectory;
-  }
-
-  @override
-  Directory getCacheArtifacts() {
-    return globals.fs.currentDirectory;
-  }
-
-  @override
-  Directory getCacheDir(String name) {
-    return globals.fs.currentDirectory;
-  }
-
-  @override
-  Directory getDownloadDir() {
-    return globals.fs.currentDirectory;
-  }
-
-  @override
-  Directory getRoot() {
-    return globals.fs.currentDirectory;
-  }
-
-  @override
-  String getHostPlatformArchName() => 'x64';
-
-  @override
-  File getLicenseFile() {
-    return globals.fs.currentDirectory.childFile('LICENSE');
-  }
-
-  @override
-  File getStampFileFor(String artifactName) {
-    throw UnsupportedError('Not supported in the fake Cache');
-  }
-
-  @override
-  String getStampFor(String artifactName) {
-    throw UnsupportedError('Not supported in the fake Cache');
-  }
-
-  @override
-  String getVersionFor(String artifactName) {
-    throw UnsupportedError('Not supported in the fake Cache');
-  }
-
-  @override
-  Directory getWebSdkDirectory() {
-    return globals.fs.currentDirectory;
-  }
-
-  @override
-  bool isOlderThanToolsStamp(FileSystemEntity entity) {
-    return false;
-  }
-
-  @override
-  Future<bool> isUpToDate() async {
-    return true;
-  }
-
-  @override
-  void setStampFor(String artifactName, String version) {
-    throw UnsupportedError('Not supported in the fake Cache');
-  }
-
-  @override
-  Future<void> updateAll(Set<DevelopmentArtifact> requiredArtifacts) async {
-  }
-
-  @override
-  Future<bool> doesRemoteExist(String message, Uri url) async {
-    return true;
-  }
-
-  @override
-  void clearStampFiles() { }
-
-  @override
-  void checkLockAcquired() { }
-
-  @override
-  Future<void> lock() async { }
-
-  @override
-  void releaseLock() { }
 }
